Emit builder Emit tasks directly + Program IR (#79)
* moving program into IR module

* fixing broken import.
